Сделал выбор фото
Пришлось сделать выбор фото не в экране редактирования т.к. не смог разобраться как через Compose навигацию перекидывать файлы. В моем случае фото могло быть Bitmap, Int, Uri. Я пробовал преобразовывать данные, но без ошибок не получалось
